requirements.txt
annotated-doc==0.0.4
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.1
async-lru==2.3.0
attrs==26.1.0
babel==2.18.0
beautifulsoup4==4.14.3
bleach==6.3.0
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cmake==4.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.20
decorator==5.2.1
defusedxml==0.7.1
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.25.2
fonttools==4.62.1
fqdn==1.5.1
fsspec==2026.3.0
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.9.1
idna==3.11
imageio==2.37.3
ipykernel==7.2.0
ipython==9.12.0
ipython-pygments-lexers==1.1.1
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
jinja2==3.1.6
joblib==1.5.3
json5==0.14.0
jsonpointer==3.1.1
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-client==8.8.0
jupyter-console==6.6.3
jupyter-core==5.9.1
jupyter-events==0.12.0
jupyter-lsp==2.3.1
jupyter-server==2.17.0
jupyter-server-terminals==0.5.4
jupyterlab==4.5.6
jupyterlab-pygments==0.3.0
jupyterlab-server==2.28.0
jupyterlab-widgets==3.0.16
kiwisolver==1.5.0
lark==1.3.1
lazy-loader==0.5
markdown-it-py==4.0.0
markupsafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mistune==3.2.0
mpmath==1.3.0
nbclient==0.10.4
nbconvert==7.17.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
notebook==7.5.5
notebook-shim==0.2.4
numpy==2.4.4
opencv-contrib-python==4.13.0.92
opencv-python==4.13.0.92
packaging==26.0
pandocfilters==1.5.1
parso==0.8.6
pillow==12.2.0
platformdirs==4.9.4
polars==1.39.3
polars-runtime-32==1.39.3
prometheus-client==0.24.1
prompt-toolkit==3.0.52
psutil==7.2.2
pure-eval==0.2.3
pycparser==3.0
pygments==2.20.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-json-logger==4.1.0
pywinpty==3.0.3
pyyaml==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.33.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-image==0.26.0
scikit-learn==1.8.0
scipy==1.17.1
send2trash==2.1.0
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
soupsieve==2.8.3
stack-data==0.6.3
sympy==1.14.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2026.3.3
timm==1.0.26
tinycss2==1.4.0
torch==2.11.0
torchvision==0.26.0
tornado==6.5.5
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.1
typing-extensions==4.15.0
tzdata==2026.1
ultralytics==8.4.34
ultralytics-thop==2.0.18
uri-template==1.3.0
urllib3==2.6.3
wcwidth==0.6.0
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
wheel==0.46.3
widgetsnbextension==4.0.15
